Override FetchContent flag to make sure that content is populated.
This flag now seems to be ON by default in our deb builds but our vendor
packages do not currently pre-populate fetchcontent contents.

I'm applying this patch at the template level assuming that even if it
does not affect Debian as-of bookworm it will in the future and is benign otherwise.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>